Sorting left

+++ Project/Code/Nearest Neighbor/struct.h
@@ -1,6 +1,6 @@
 struct Nodes {
    int x;
    int y;
-   int *neighbourDistance;
+   double *neighbourDistance;
  
 };
